+++ src/main/java/dev/knapp/security/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package dev.knapp.security;
 
-import com.revature.p2.beans.User;
+import dev.knapp.beans.User;
 import io.jsonwebtoken.*;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
@@ -9,8 +9,8 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.revature.p2.security.SecurityConstants.EXPIRATION_TIME;
-import static com.revature.p2.security.SecurityConstants.SECRET;
+import static dev.knapp.security.SecurityConstants.EXPIRATION_TIME;
+import static dev.knapp.security.SecurityConstants.SECRET;
 
 @Component
 public class JwtTokenProvider {
@@ -20,9 +20,9 @@ public class JwtTokenProvider {
 
         Date now = new Date(System.currentTimeMillis());
         Date expiryDate = new Date(now.getTime()+EXPIRATION_TIME);
-        String userId = Long.toString(user.getId());
+        String userId = Integer.toString(user.getId());
         Map<String, Object> claims = new HashMap<>();
-        claims.put("id", (Long.toString(user.getId())));
+        claims.put("id", (Integer.toString(user.getId())));
         claims.put("username", user.getUsername());
 //        claims.put("fullName", user.getFullName());
         return Jwts.builder()
@@ -53,10 +53,10 @@ public class JwtTokenProvider {
     }
 
     //Get user Id from Token
-    public Long getUserIdFromJWT(String token){
+    public Integer getUserIdFromJWT(String token){
         Claims claims = Jwts.parser().setSigningKey(SECRET).parseClaimsJws(token).getBody();
         String id = (String) claims.get("id");
-        return Long.parseLong(id);
+        return Integer.parseInt(id);
 
     }
 }
